% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_to_se.R
\name{ci_to_se}
\alias{ci_to_se}
\title{Calculate standard error based on confidence intervals}
\usage{
ci_to_se(lower, upper)
}
\arguments{
\item{lower}{A lower confidence interval bound}

\item{upper}{An upper confidence interval bound}
}
\value{
A standard error value
}
\description{
Calculate standard error based on confidence intervals
}
\examples{
ci_to_se(-0.12, 1.71)
ci_to_se(0.07, 1.12)
}
