% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firepower.R
\name{firepower}
\alias{firepower}
\title{Calculate the median study power for a meta-analysis and visualise this data using a Firepower plot}
\usage{
firepower(power_list)
}
\arguments{
\item{power_list}{A list of dataframes with calculated power, derived from the "mapower_ul" or "mapower_se"
functions. See 'Examples' for how to make a list.}
}
\value{
This function returns the following:
      \item{dat}{A dataset with the median results from power analyses for a range of effect sizes, including the
      specified true effect size, in a column labelled "power_true". The additional added columns include
      results for power analysis assuming a range of true effect sizes, beginning at 0.1 ("power_es01"),
      then 0.2 ("power_es02"), then continuing in increments of 0.1 up to 1 ("power_es1").}
      \item{fp_plot}{A firepower plot}
}
\description{
Creates a dataset with median power and a firepower plot, which visualises the median power for the
assuming the observed effect size in the meta-analysis is the true effect size and assuming a
range of true effect sizes ranging from 0.1 to 1 in increments of 0.1. See 'Details'.
}
\details{
\strong{The true effect size}

It is usually difficult to establish a "true" effect size in practice, as reported effect sizes are typically
inflated. Firepower plots implements two alternative categories for a true effect for the purposes of power
analysis: The observed summary effect size estimate reported in the meta-analysisand statistical power for
a range of true effect sizes (0.1 to 1, in increments of 0.1).
}
\examples{
### Calcuate median power for meta-analyses
power_ooi <- mapower_se(dat = dat_ooi, observed_es = 0.178, name = "ooi et al 2017")
power_med_ooi <- power_ooi$power_median_dat

keech_power <- mapower_ul(dat = dat_keech, observed_es = 0.08, name = "Keech et al 2017")
power_med_keech <- keech_power$power_median_dat

power_bakermans_kranenburg <- mapower_se(dat = dat_bakermans_kranenburg, observed_es = 0.32, name = "Bakermans-Kranenburg et al 2013")
power_med_bakermans_kranenburg <- power_bakermans_kranenburg$power_median_dat

### Create a list
list_power <- list(power_med_ooi, power_med_keech, power_med_bakermans_kranenburg)

### Run firepower function
fp <- firepower(list_power)

### Extract data from datasets in list
power_dat <- fp$dat
power_dat

### Create firepower plot
fp_plot <- fp$fp_plot
fp_plot
}
